Create Data Logger for New PZEM004T
This project arose from the need to record the energy data of the new PZEM! I already had the old version of PZEM004T, but when buying a new one, I came across the hardware change and need to change the program! Nothing worked until the library was kindly provided by our colleague Mandulaj (PZEM-004T-v30) to whom I am very grateful! Link to library: https://github.com/mandulaj/PZEM-004T-v30 The project consists of an Arduino Mega 2560, an RTC1307, an IC2 20x4 LCD display and a microSD reader. Data is recorded at 30-second intervals in a file with the name formed by the date! Changing the date automatically starts a new file!